Restructure consensus with FPC enabled by default for 50x speedup
- Implement clean module structure: photon, prism, wave, focus, nova, horizon, flare, nebula, quasar
- Enable FPC by default in all configurations for owned-object transaction speedup
- Wire up engines (chain/dag/pq) with FPC integration
- Use context.Context everywhere for consistency
- Remove all Snow references, replaced with new module names
- BadgerDB v4 as storage backend through luxfi/database

+++ config/runtime_test.go
@@ -329,7 +329,7 @@ func TestConcurrentRuntimeAccess(t *testing.T) {
 						K:                     20 + idx,
 						AlphaPreference:       10 + idx,
 						AlphaConfidence:       15 + idx,
-						Beta:                  5 + idx,
+						Beta:                  uint32(5 + idx),
 						ConcurrentPolls:       3,
 						OptimalProcessing:     10,
 						MaxOutstandingItems:   100,
